// frontend/src/generated/model/leaderboardData.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * AnthoLume API v1
 * REST API for AnthoLume document management system
 * OpenAPI spec version: 1.0.0
 */
import type { LeaderboardEntry } from './leaderboardEntry';

export interface LeaderboardData {
  all: LeaderboardEntry[];
  year: LeaderboardEntry[];
  month: LeaderboardEntry[];
  week: LeaderboardEntry[];
}