// frontend/src/generated/model/userStreak.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * AnthoLume API v1
 * REST API for AnthoLume document management system
 * OpenAPI spec version: 1.0.0
 */

export interface UserStreak {
  window: string;
  max_streak: number;
  max_streak_start_date: string;
  max_streak_end_date: string;
  current_streak: number;
  current_streak_start_date: string;
  current_streak_end_date: string;
}